some changes to the api

+++ src/element.rs
@@ -11,7 +11,7 @@ pub struct Bound {
     pub dim: Vec2<u32>,
     pub offset: Vec2<f32>,
     pub overflow: Overflow,
-    pub shadow: [f32; 4], // native representation for padding [left, right, top, low]
+    pub shadow: [u32; 4], // native representation for padding [left, right, top, low]
     pub angle: f32,
 }
 
@@ -22,7 +22,7 @@ impl Bound {
             pos: Vec2::new(0, 0),
             offset: Vec2::new(0.0, 0.0),
             overflow: Overflow::Clip { active: false },
-            shadow: [0.0, 0.0, 0.0, 0.0],
+            shadow: [0, 0, 0, 0],
             angle: 0.0,
         }
     }
